Move submissionDatWithLabels helper to presenter folder
