+++ lib/controller/page/type/page.summary/page.summary.controller.js
@@ -24,7 +24,7 @@ const {getUrl} = require('../../../../route/route')
 const redirectNextPage = require('../../../../page/redirect-next-page/redirect-next-page')
 const {checkSubmits} = require('../../../../page/check-submits/check-submits')
 const {pdfPayload} = require('../../../../presenter/pdf-payload')
-const {submissionDataWithLabels} = require('../../../../middleware/routes-output/submission-data-with-labels')
+const {submissionDataWithLabels} = require('../../../../presenter/submission-data-with-labels')
 
 const {format} = require('../../../../format/format')
