+++ lib/controller/page/type/page.summary/page.summary.controller.unit.spec.js
@@ -30,7 +30,7 @@ const pageSummaryController = proxyquire('./page.summary.controller', {
   '../../../../page/redirect-next-page/redirect-next-page': redirectNextPageStub,
   '../../../../client/submitter/submitter': {submitterClient: submitterClientSpy},
   '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-  '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub}
+  '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub}
 })
 
 getInstancePropertyStub.callsFake((_id, type) => {
@@ -170,7 +170,7 @@ test('it attaches a user submission when the property is set to true', async t =
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
     '../../../../constants/constants': {
       SERVICE_OUTPUT_EMAIL: 'bob@gov.uk'
     }
@@ -197,7 +197,7 @@ test('it does not attach a user submission when the property is set to false', a
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
     '../../../../constants/constants': {
       SERVICE_OUTPUT_EMAIL: 'bob@gov.uk'
     }
@@ -219,7 +219,7 @@ test('it attaches json to submission when env var present', async t => {
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
     '../../../../constants/constants': {
       SERVICE_OUTPUT_JSON_ENDPOINT: 'https://example.com/adaptor',
       SERVICE_OUTPUT_JSON_KEY: 'shared_key',
@@ -247,7 +247,7 @@ test('when there are files it attaches them to json submission', async t => {
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
     '../../../../constants/constants': {
       SERVICE_OUTPUT_JSON_ENDPOINT: 'https://example.com/adaptor',
       SERVICE_OUTPUT_JSON_KEY: 'shared_key',
@@ -281,7 +281,7 @@ test('when there are files it attaches to the root key attachments', async t => 
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub}
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub}
   })
 
   const attachments = [
@@ -315,7 +315,7 @@ test('when there are files it attaches to the root key attachments', async t => 
 test('it does not attach json to submission when env var not present', async t => {
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../constants/constants': {
       SERVICE_OUTPUT_EMAIL: 'bob@gov.uk'
@@ -339,7 +339,7 @@ test('Dynamic content is rendered in the Team email body', async t => {
 
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../constants/constants': {
       SERVICE_OUTPUT_EMAIL: 'bob@gov.uk'
@@ -366,7 +366,7 @@ test('user email body can be overridden to use a rendered template', async t => 
 
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../constants/constants': {
       SERVICE_OUTPUT_EMAIL: 'bob@gov.uk'
@@ -399,7 +399,7 @@ test('actions contains email and json actions', async t => {
 
   const pageSummaryController = proxyquire('./page.summary.controller', {
     '../../../../presenter/pdf-payload': {pdfPayload: pdfPayloadStub},
-    '../../../../middleware/routes-output/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
+    '../../../../presenter/submission-data-with-labels': {submissionDataWithLabels: submissionDataWithLabelsStub},
     '../../../../page/check-submits/check-submits': {checkSubmits: checkSubmitsStub},
     '../../../../constants/constants': {
       SERVICE_OUTPUT_JSON_ENDPOINT: 'https://example.com/adaptor',
