+++ lib/presenter/submission-data-with-labels.js
@@ -1,5 +1,5 @@
-const {getInstanceController} = require('../../controller/controller')
-const {formatProperties} = require('../../page/page')
+const {getInstanceController} = require('../controller/controller')
+const {formatProperties} = require('../page/page')
 
 const submissionDataWithLabels = (title, heading, subHeading, userData) => {
   const answersComponent = {
